♻️  refactor : PopOver 테두리 분리
