+++ src/components/Common/AddContentPopOver/AddContentPopOver.module.scss
@@ -1,0 +1,15 @@
+.container {
+  p {
+    height: 40px;
+    padding: 0 10px;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    list-style: none;
+    font-size: 1rem;
+    cursor: pointer;
+  }
+  p:hover {
+    background-color: #f9f9f9;
+  }
+}
