--- src/components/Common/PopOver/PopOver.module.scss
@@ -1,29 +1,0 @@
-.container {
-  width: 230px;
-  margin: 0;
-  padding: 20px 0;
-  position: absolute;
-  border: 1px solid #dbdbdb;
-  font-size: 1.4rem;
-  bottom: 70px;
-  left: 90px;
-  border-radius: 8px;
-
-  .content-list {
-    height: 40px;
-    padding: 0 10px;
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    list-style: none;
-    cursor: pointer;
-  }
-
-  li:hover {
-    background-color: #f9f9f9;
-  }
-
-  @media (min-width: 768px) {
-    bottom: 0px;
-  }
-}
