+++ src/pages/index.page.tsx
@@ -1,3 +1,3 @@
 export default function Home() {
-  return <div>Home</div>;
+  return <div>home</div>;
 }
